requirements.txt
boto3
dagit
dagster
dagster-docker
dagster-gcp
dagster-gcp-pandas
dagster-postgres
dagster-webserver
dagster-cloud
duckdb
google-auth
google-cloud-bigquery
Jinja2
matplotlib
#oscrypto # for now, install from gh to fix this: https://community.snowflake.com/s/article/Python-Connector-fails-to-connect-with-LibraryNotFoundError-Error-detecting-the-version-of-libcrypto
git+https://github.com/wbond/oscrypto.git@master#egg=oscrypto
openai
pandas
pandas-gbq
plotly
pyod
pyyaml
snowflake-connector-python[pandas]
seaborn
streamlit
tabulate
